+++ miles/utils/tracking_utils.py
@@ -1,37 +1,19 @@
 import logging
 
-import wandb
-from miles.utils.tensorboard_utils import _TensorboardAdapter
-
-from . import wandb_utils
-from .prometheus_utils import get_prometheus, init_prometheus
+from .tracking import TrackingManager
 
 logger = logging.getLogger(__name__)
+_manager = TrackingManager()
 
 
 def init_tracking(args, primary: bool = True, **kwargs):
-    if primary:
-        wandb_utils.init_wandb_primary(args, **kwargs)
-    else:
-        wandb_utils.init_wandb_secondary(args, **kwargs)
-
-    if args.use_prometheus:
-        init_prometheus(args, start_server=primary)
+    _manager.init(args, primary=primary, **kwargs)
 
 
-# TODO further refactor, e.g. put TensorBoard init to the "init" part
 def log(args, metrics, step_key: str):
-    if args.use_wandb:
-        wandb.log(metrics)
+    step = metrics.get(step_key)
+    _manager.log(metrics, step=step)
 
-    if args.use_tensorboard:
-        metrics_except_step = {k: v for k, v in metrics.items() if k != step_key}
-        _TensorboardAdapter(args).log(data=metrics_except_step, step=metrics[step_key])
 
-    if args.use_prometheus:
-        prom = get_prometheus()
-        assert prom is not None, (
-            "Prometheus collector is not initialized; ensure init_tracking(..., primary=...) ran on the "
-            "driver and workers can resolve the miles_prometheus_collector Ray actor."
-        )
-        prom.update.remote(metrics)
+def finish_tracking():
+    _manager.finish()
