+++ train.py
@@ -7,7 +7,7 @@ from miles.utils.arguments import parse_args
 from miles.utils.async_utils import eager_create_task
 from miles.utils.logging_utils import configure_logger
 from miles.utils.misc import should_run_periodic_action
-from miles.utils.tracking_utils import init_tracking
+from miles.utils.tracking_utils import finish_tracking, init_tracking
 
 
 async def train(args):
@@ -106,4 +106,7 @@ async def train(args):
 
 if __name__ == "__main__":
     args = parse_args()
-    asyncio.run(train(args))
+    try:
+        asyncio.run(train(args))
+    finally:
+        finish_tracking()
